+++ src/provider/ui/components/ButtonNotification.tsx
@@ -1,13 +1,14 @@
 import * as React from 'react';
 import { NotificationTime } from './NotificationTime';
-import { INotificationProps } from '../../models/INotificationProps';
-import {NotificationCenterAPI} from '../../NotificationCenterAPI';
+import { Button } from './Button';
+import { INotificationProps } from '../models/INotificationProps';
+import {NotificationCenterAPI} from '../NotificationCenterAPI';
 declare var window: Window&{openfin: {notifications: NotificationCenterAPI}};
 
 /**
- * Displays a single notification within the UI
+ * Displays a button notification within the UI
  */
-export class Notification extends React.Component<INotificationProps, {}> {
+export class ButtonNotification extends React.Component<INotificationProps, {}> {
     private handleNotificationClose(e: React.MouseEvent<HTMLElement>) {
         e.stopPropagation();
         e.nativeEvent.stopImmediatePropagation();
@@ -15,6 +16,19 @@ export class Notification extends React.Component<INotificationProps, {}> {
     }
 
     public render(): React.ReactNode {
+        let buttons = null;
+        if (this.props.meta.buttons) {
+            buttons = this.props.meta.buttons.map((button, idx) => {
+                return (
+                    <Button
+                        key={idx}
+                        buttonIndex={idx}
+                        meta={this.props.meta}
+                    />
+                );
+            });
+        }
+
         return (
             <li
                 className="notification-item"
@@ -26,7 +40,7 @@ export class Notification extends React.Component<INotificationProps, {}> {
                     alt=""
                     onClick={e => this.handleNotificationClose(e)}
                 />
-                <NotificationTime date={this.props.meta.date} />
+                <NotificationTime date={new Date(this.props.meta.date)} />
                 <div className="notification-body">
                     <div className="notification-source">
                         <img
@@ -43,6 +57,7 @@ export class Notification extends React.Component<INotificationProps, {}> {
                     <div className="notification-body-text">
                         {this.props.meta.body}
                     </div>
+                    <div id="notification-body-buttons">{buttons}</div>
                 </div>
             </li>
         );
