+++ src/provider/ui/components/Notification/Notification.tsx
@@ -1,14 +1,13 @@
 import * as React from 'react';
-import { NotificationTime } from './NotificationTime';
-import { Button } from './Button';
+import { NotificationTime } from './NotificationTime/NotificationTime';
 import { INotificationProps } from '../../models/INotificationProps';
 import {NotificationCenterAPI} from '../../NotificationCenterAPI';
 declare var window: Window&{openfin: {notifications: NotificationCenterAPI}};
 
 /**
- * Displays a button notification within the UI
+ * Displays a single notification within the UI
  */
-export class ButtonNotification extends React.Component<INotificationProps, {}> {
+export class Notification extends React.Component<INotificationProps, {}> {
     private handleNotificationClose(e: React.MouseEvent<HTMLElement>) {
         e.stopPropagation();
         e.nativeEvent.stopImmediatePropagation();
@@ -16,19 +15,6 @@ export class ButtonNotification extends React.Component<INotificationProps, {}> 
     }
 
     public render(): React.ReactNode {
-        let buttons = null;
-        if (this.props.meta.buttons) {
-            buttons = this.props.meta.buttons.map((button, idx) => {
-                return (
-                    <Button
-                        key={idx}
-                        buttonIndex={idx}
-                        meta={this.props.meta}
-                    />
-                );
-            });
-        }
-
         return (
             <li
                 className="notification-item"
@@ -40,7 +26,7 @@ export class ButtonNotification extends React.Component<INotificationProps, {}> 
                     alt=""
                     onClick={e => this.handleNotificationClose(e)}
                 />
-                <NotificationTime date={new Date(this.props.meta.date)} />
+                <NotificationTime date={this.props.meta.date} />
                 <div className="notification-body">
                     <div className="notification-source">
                         <img
@@ -57,7 +43,6 @@ export class ButtonNotification extends React.Component<INotificationProps, {}> 
                     <div className="notification-body-text">
                         {this.props.meta.body}
                     </div>
-                    <div id="notification-body-buttons">{buttons}</div>
                 </div>
             </li>
         );
