+++ src/provider/ui/container/App.tsx
@@ -1,7 +1,7 @@
-import * as React from 'react';
-import { Header } from './Header';
-import { NotificationView } from './NotificationView';
-import { Footer } from './Footer';
+import * as React  from 'react';
+import { Header } from '../components/Header/Header';
+import { NotificationView } from '../components/NotificationView/NotificationView';
+import { Footer } from '../components/Footer/Footer';
 
 interface IAppState {
     groupBy?: eGroupMethod;
