+++ src/provider/ui/js/index.tsx
@@ -1,6 +1,6 @@
 import* as React from 'react';
 import* as ReactDOM from 'react-dom';
-import{App} from '../components/App';
+import{App} from '../container/App';
 
 import '../../../../res/provider/ui/css/launch-bar.css';
 import '../../../../res/provider/ui/css/notifications.css';
